change file naming to have same file as feature/34/endpoint-screenshots.
