--- src/server/api/routes/screenshot.router.js
@@ -1,25 +1,0 @@
-"use strict";
-
-// router setup
-const express = require("express");
-const router = express.Router({ mergeParams: true });
-
-// controllers
-const screenshotController = require("../controllers/screenshot.controller");
-
-// ENDPOINT: /api/screenshots/ :POST
-router.post("/", (req, res, next) => {
-  screenshotController
-    .createScreenshot(req.body)
-    .then(result => res.json(result))
-    .catch(error => {
-      console.log(error);
-
-      res
-        .status(400)
-        .send("Bad request")
-        .end();
-    });
-});
-
-module.exports = router;
